requirements.txt
customtkinter
aliyun-python-sdk-core-v3
aliyun-python-sdk-cdn
python-dotenv
